requirements.txt
alembic==1.8.1
anyio==3.6.1
boto3==1.24.91
botocore==1.27.91
click==8.1.3
colorama==0.4.5
ecdsa==0.18.0
fastapi==0.85.0
greenlet==1.1.3.post0
h11==0.14.0
httptools==0.5.0
idna==3.4
jmespath==1.0.1
Mako==1.2.3
MarkupSafe==2.1.1
numpy==1.23.4
pandas==1.5.0
passlib==1.7.4
psycopg2==2.9.4
pyasn1==0.4.8
pydantic==1.10.2
python-dateutil==2.8.2
python-dotenv==0.21.0
python-jose==3.3.0
python-multipart==0.0.5
pytz==2022.4
PyYAML==6.0
rsa==4.9
s3transfer==0.6.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.41
starlette==0.20.4
typing_extensions==4.4.0
urllib3==1.26.12
uvicorn==0.18.3
watchfiles==0.17.0
websockets==10.3
